Improvements and fixes proposed by Fatima
git-svn-id: https://svn.squada.com/svn/consulting30/trunk@415 92e3e9a8-b16a-40ff-a0f0-73f7c07c1316

+++ web/src/consulting/forms.py
@@ -71,7 +71,7 @@ class MedicineForm(forms.ModelForm):
                                     a los síntomas'),
                                     choices=BEFORE_AFTER_CHOICES,
                                     validators=[validate_choice])
-    months = forms.IntegerField(label=_(u'Número de meses de toma del\
+    months = forms.DecimalField(label=_(u'Número de meses de toma del\
                                          fármaco'))
     posology = forms.IntegerField(label=_(u'Posología (mg/día)'))
